Rename namespace for data-app dev things
