+++ enterprise/frontend/src/embedding-sdk-package/package.template.json
@@ -24,10 +24,10 @@
       "require": "./dist/data-app.cjs",
       "import": "./dist/data-app.js"
     },
-    "./data-app-sandbox": {
-      "types": "./dist/data-app-sandbox.d.ts",
-      "require": "./dist/data-app-sandbox.cjs",
-      "import": "./dist/data-app-sandbox.js"
+    "./data-app-dev": {
+      "types": "./dist/data-app-dev.d.ts",
+      "require": "./dist/data-app-dev.cjs",
+      "import": "./dist/data-app-dev.js"
     },
     "./package.json": "./package.json"
   },
